Reboot the project in Haskell instead of Ocaml
The original plan in OCaml was to make the Genlex module to all the
heavy lifting for parsing the strace output, however this didn't work
out because it can't be convinced to parse C-style hex literals, and
it was too much hassle to keep track of line-endings. Although the
ocamllex and ocamlyacc tools are quite good, they are more heavyweight
than I originally intended.

Use Haskell instead so that we can use the excellent Parsec module to
parse more flexibly, but with less hassle. Haskell will also probably
make the task of interpreting the trace a little easier due to
laziness.

Also, this project is for fun, and I find Haskell more fun these days.